#173 ViewModel 변경에 따른 코드 변경
